backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
opencv-python==4.10.0.84
mediapipe==0.10.14
ultralytics==8.2.0
numpy==1.26.4
Pillow==10.4.0
pydantic==2.8.2
pydantic-settings==2.4.0
redis==5.0.8
slowapi==0.1.9
aiofiles==24.1.0
